+++ eslint.config.js
@@ -1,6 +1,5 @@
 import * as dovenv from './packages/config/eslint-config/src/main.mjs'
 
-// console.dir( dovenv.generalConfig, { depth: 4 } )
 export default [
 	dovenv.includeGitIgnore(),
 	...dovenv.config,