+++ packages/plugin/repo/src/gh/_super.ts
@@ -2,9 +2,21 @@ import { execChild } from '@dovenv/core/utils'
 
 import { Repo } from '../_super/main'
 
+type Repository = {
+	url   : string
+	name  : string
+	owner: {
+		id    : string
+		login : string
+	}
+	description?      : string
+	homepageUrl?      : string
+	repositoryTopics? : { name: string }[]
+}
+
 export class GHSuper extends Repo {
 
-	async getRepositories( opts?: {
+	async getRepoList( opts?: {
 		archived?   : boolean
 		fork?       : boolean
 		visibility? : 'public' | 'private' | 'internal'
@@ -27,10 +39,21 @@ export class GHSuper extends Repo {
 		else if ( fork === false ) flags.push( '--source' )
 
 		if ( visibility ) flags.push( `--visibility ${visibility}` )
+		const command = `gh repo list ${this.opts?.userID || ''} --json url,name,owner,homepageUrl,description,repositoryTopics`
+		const {
+			stdout, stderr,
+		} = await execChild( flags.length ? `${command} ${flags.join( ' ' )}` : command )
+		if ( stderr ) throw new Error( `Error getting repo list: ${stderr}` )
+		const repos = JSON.parse( stdout.trim() ) as Repository[]
 
-		const { stdout } = await execChild( flags.length ? `gh repo list ${flags.join( ' ' )}` : `gh repo list` )
-
-		return ' ' + stdout.trim()
+		return repos.map( r => ( {
+			desc     : r.description?.trim() === '' ? undefined : r.description,
+			homepage : r.homepageUrl?.trim() === '' ? undefined : r.homepageUrl,
+			name     : r.name,
+			owner    : r.owner.login,
+			topics   : r.repositoryTopics?.map( t => t.name ),
+			url      : r.url,
+		} ) )
 
 	}