+++ packages/core/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @dovenv/core
 
+## 1.0.1
+
+### Patch Changes
+
+- fix create plugin
+
+- Updated dependencies []:
+  - @dovenv/utils@1.0.1
+
+## 0.2.0
+
+### Minor Changes
+
+- fix beta update
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dovenv/utils@0.2.0
+
+## 0.1.0
+
+### Minor Changes
+
+- update to beta
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dovenv/utils@0.1.0
+
 ## 0.0.20
 
 ### Patch Changes